Refactor MONITOR_CLASS_NAMES to be used w/o gems
MiqServer::WorkerManagement::Monitor::ClassNames is a module nested
inside of the MiqServer's namespace, making it nearly impossible to load
it on it's own without including the rest of the classes and subclasses
that are included by it.

Moving the constants into a global constant then allows the
MiqServer::WorkerManagement::Monitor::ClassNames model to reference
the constants defined there, and allows it to be used by plain ruby
without the overhead of rails/activesupport/etc.