Do not cache realtime at this point
We need to refactor a lot of surrounding code, before we can cache
realtime here.

The problem is that we create dozens of MiqTask for each
perf_capture_timer run and we keep all the MiqTask numbers within the
miq_callback of realtimes. Therefore -> we always update realtime, even
though, we should not. That's tmrw's task.