Fix erratic spec for reports with custom attributes
`Consumption#consumed_hours_in_interval` regards the consumption start
as the greater of the passed in `@start_time`, or `born_at`, which
will be the timestamp of the first rollup for the resource. If this
test is run any time in the 23rd hour UTC, the consumption period will
be < 1 hour, causing this to return `0`, failing the test.

By creating the rollup at the start of day UTC, we avoid this edge
case, ensuring that this test passes at any time of day.