Workaround Rails.configuration.database_configuration being {}
If you don't have a config/database.yml and only ENV["DATABASE_URL"],
Rails.configuration.database_configuration will be {}, so you'll get an
undefined [] for nil error.

While attempting to fix that in https://github.com/rails/rails/pull/29305,
I was told that it was not intended to be the public interface.  So I'm
changing this PR to always use ActiveRecord::Base.configurations instead.