This comment is useless
There are operations like intersection and union in the algorithm.
But these operations are not same as we know from maths.
They can recieve also nil as operand and it means:
don't use this operand and don't apply operator here.

For example mentioned operators works in this way:

[1] INTERSECTION [1, 2] = [1]
[] INTERSECTION [1, 2] = []
nil INTERSECTION [1, 2] = [1, 2]

operands represents filter as array of ids,
when filter is not array but nil
it means don't apply the filter