Setup namespace in loggers before child requires
When doing the following:

  Dir.glob(File.join(File.dirname(__FILE__), "loggers", "*")).each { |f| require f }

We are including a bunch of classes do the following:

  module Vmdb::Loggers
    class MyLogger < VmdbLogger
    end
  end

This works fine when the module Vmdb::Loggers has been defined, but when
requiring `lib/vmdb/loggers.rb` directly, this is usually not the case.
By pushing the requires of the child loggers to the bottom of the file,
we ensure that the `Vmdb` module has been defined, and it's child,
`Loggers`, has also been defined.

This allows the following:

  require "vmdb/loggers"
  require "vmdb/logging"

To work with minimal prerequisites and doesn't require autoloading from
rails.