Separate arel generation from field/tag parsing
When `Field` was extracted it was attractive to let it be a middleman
between `MiqExpression` and the Arel library since everything goes
through `Field#arel_attribute`. While this started out as
straightforward delegation, more supporting code was required to
bridge more complex expressions such as `CONTAINS`, or things that
require extra arguments like `matches`, etc..

I think that consequently `Field` (and `Tag`) have taken too much on
in terms of responsibility. It also introduced a cohesion problem in
that not all fields are concerned with columns in the database.

While this change might be a step back for `MiqExpression`, I think it
greatly improves `Field`. I think there are better patterns out there
for better separating responsibilities here, but that can be left to a
future revision.