Make Vmdb::Settings::Walker a independent module
The code in Vmdb::Settings::Walker was simply vanilla hash traversal
code that was used to modify the values found in the settings, or any
other settings like hash (usually password decryption/encryption).

This code was also used independently of the of the main Vmdb::Settings
class, in things like the `fix_auth` tool, but when loaded prior to the
Vmdb::Settings, if autoloading was enabled, the main class would never
be loaded because the empty class is technically defined in the
`lib/vmdb/settings/walker.rb` file.

The changes done here are as follows:

* Changes the Vmdb::Settings::Walker to be a nested module only (no
  class definitions).  While the file path hasn't changed, the module is
  now Vmdb:SettingsWalker (slightly breaks normal conventions)
* Changes the method definitions to be within a `ClassMethods` module
  that are then extended by the outer module to allow it to behave the
  same as it had (a module with top level class methods that could be
  called without needing Vmdb::Settings).
* Update all callers to use the changed namespace instead, and extend
  where needed
* In Vmdb::Settings specifically, removed duplicate method definitions
  now that extending the the ClassMethods module defined the same
  methods