Let queue workers process an active message
https://bugzilla.redhat.com/show_bug.cgi?id=1481800

In e5f4bd3fe1299070e40d235be963428b4f9a2d14, we added a 10 minute
timeout that would give workers a little time to complete their work
after they exceed their memory threshold before we'd kill them.  This
causes workers to be killed prematurely before completing the work item.

What we really want is for the work item to complete but kill the worker
if the worker has exceeded memory/time thresholds and the work item hasn't
completed in a reasonable time.  This reasonable time is the msg_timeout
associated with the queue message.