Add an example Procfile and a rake task for starting the app using foreman
This allows developers to start workers independently of the server
in a very easy and repeatable way.

The rake task takes care of ensuring the "server" process looks started
so that the workers will run normally. Additional roles need to be
set manually as the server process will not be running.

The intended use is for a developer to create their own Procfile
in the root of the repo with whatever workers they need to start
for the work that they are doing. Procfile was added to .gitignore
so that changes to the Procfile don't dirty the git status.