Return VMs and Templates for EMS prev_relats
When retrieving vmdb_relats_ems we weren't including templates in the
list of children for folders.  This isn't an issue for clusters, hosts,
and resource_pools because templates don't belong to any of these but
they do belong to folders.

If a VM was marked as a template and moved into a new folder it was not
being included in the prev_relats which meant that its previous
resource_pool and folder relat were not being deleted, but we were
adding a new folder relationship.  This caused the "multiple parents
found" exception seen occasionally on the UI while provisioning.

https://bugzilla.redhat.com/show_bug.cgi?id=1475405