Set user's group to the requester group.
User's current group might have changed before the provision finishes.
So the user's current group from DB may be different from the user's group when the request is made.

Need to set the user's group back to requester group the user is in when the provision is submitted.
Therefore the provisioned instances may belong to the right user/group/tenant.

https://bugzilla.redhat.com/show_bug.cgi?id=1467364