Fail with descriptive message when no EMS
When no EMS is associated with a VM or template, the provision request
fails with a non-descriptive message. Such case might happen when the VM
or template were removed from the provider side, but still exist as
archived on vmdb. The purpose of the PR is to provide a clear failure
message when there is no EMS.

https://bugzilla.redhat.com/show_bug.cgi?id=1471124