Add DockerEmbeddedAnsible
This will run the containers which make up AWX (https://github.com/ansible/awx)
and configure our app to use that for the embedded ansible feature.

This class uses the docker-api gem to communicate with the locally
running docker daemon to pull and launch the containers.

We use port 54321 as the host port so that this can be used
seamlessly in place of ApplianceEmbeddedAnsible when ansible tower
is not installed locally