Use in_my_region when looking up group during widget content generation
Ensure that when an miq group is looked up by its description during widget content generation that in_my_region is used. Currently it is causing an error when generating content for a user because it sets the correct user but with a group in another zone, causing it to be unviewable.

https://bugzilla.redhat.com/show_bug.cgi?id=1505241