Replace ContainerLabelTagMapping.controls_tag? with a prefix-based scope
A scope is needed for tag mapping in graph refresh, and may improve
performance.
Alternative to #16449 - this one no longer cares if there exists a
ContainerLabelTagMapping linked to the category, only that it's name
starts with `kubernetes:` / `amazon:`.  This can happen to
user-created categories too, so we also check it's read_only.