Rescue JSON::ParserError in DockerEmbeddedAnsible.alive?
This error will be raised when the containers are just started.
Every API end point during the initial migration will return an
html page rather than a json payload.

This accounts for that specific situation by assuming if we don't
get a valid json response the service is not ready to serve requests