Catalog Item type list is dependent on installed providers
https://bugzilla.redhat.com/show_bug.cgi?id=1515371

Currently we display all catalog item types irrespective of what
providers are installed in the appliance. This PR links the catalog
items to different providers and sets the display attribute based
on the presence of specific provider(s). This would be used by the
UI to list the valid catalog item types