Use LinuxAdmin::SubscriptionManager directly.
LinuxAdmin::RegistrationSystem can't auto-detect the type when using
SubscriptionManager with a proxy due to its need to use the proxy info to
validate that it is registered.

Since RHN is no longer supported, we can use the SubscriptionManager class
directly.

Also, pass the registration options (proxy info) to the registered? method
since it is required when using a proxy.

Add linux_admin to the Gemfile since we use it directly

https://bugzilla.redhat.com/show_bug.cgi?id=1518695