Separate the logic for initializing a worker record and saving it
This will allow the worker (in the container) to create its own
GUID. This is necessary for worker types with multiple replicas
because we cannot tell an individual replica what its GUID is
supposed to be.

Also, we need the queue_name parameter for the deployment
per worker model. In this case we want to be able to reference
the queue name on the worker instance, but still don't want
to save that instance to the database.

We can't use create_worker_record as is, because run_single_worker.rb
also needs a way to save the worker record when it's running in a container