Scan Storage via EMS not Host
Storages are currently scanned via a host which the user has added
valid credentials for.  Storages are looked up by name which can be
different between the vCenter (where the refresh is done) and the ESXi
Host (where the scan is done).  If the name is different the scan will
fail with the following error:
`[MiqException::MiqVimResourceNotFound]: Could not find datastore: NFS-Datastore-1`

Since datastore files are scanned via the VIM SDK there is no
requirement on running the scan through the host.  We can scan the
datastore through the EMS and remove the restriction on needing to add
host credentials to be able to run storage smartstate.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1529725