Server shuts down if memory usage exceeds a limit
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1533484

Previously, the server could slowly grow in memory usage and nothing was
monitoring this.  Now, we use the memory usage we already were
collecting and react to it by shutting down the server process if it
exceeds the configurable threshold.  We rely upon systemd, or kubernetes
in the future, to see our non-zero exit code and automatically start a
new server process within a few minutes.

Note, we are trying to be "graceful" in that we call shutdown_and_exit
with a non-zero exit code.  This will deactivate the roles on the
server and attempt to gracefully shutdown the workers.

Additionally, we create a notification so an administrator can look
into it and report any future leaks.