Fix sporadic VmdbDatabaseLock test failure
If the wait_for_lock thread doesn't get to the transaction before the
loop waiting for a connection to be blocked_by another this will always
fail because the connection list is never reloaded.

This fixes the connection list reload as well as the buggy loop waiting
for the thread to be blocked on the database execution.